New stateful impl
