+++ crates/kobold/src/lib.rs
@@ -296,8 +296,6 @@ mod value;
 #[cfg(feature = "stateful")]
 pub mod stateful;
 
-pub mod stateful_new;
-
 /// The prelude module with most commonly used types.
 ///
 /// Intended use is:
